pyproject.toml
[tool.poetry]
name = "datapulse-be"
version = "0.1.0"
description = "A comprehensive cloud data management sphere."
authors = ["Faakhir30 <zahidfaakhir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-ldap = "^3.4.4"
pyjwt = "^2.9.0"
fastapi = {extras = ["all"], version = "^0.115.4"}
ldap3 = "^2.9.1"
psycopg2-binary = "^2.9.10"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
